requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.1.0
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
black==26.1.0
bleach==6.3.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docx2pdf==0.1.8
executing==2.2.1
fastjsonschema==2.21.2
fqdn==1.5.1
frozenlist==1.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.1.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
isort==7.0.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
lark==1.3.1
librt==0.7.8
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.2.0
multidict==6.7.1
mypy==1.19.1
mypy_extensions==1.1.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
openai==2.16.0
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pathspec==1.0.4
platformdirs==4.5.1
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pure_eval==0.2.3
pyaml==25.7.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytokens==0.4.1
pywin32==311
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
ruff==0.15.0
Send2Trash==2.1.0
setuptools==80.10.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
wcwidth==0.5.3
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
xai-sdk==1.6.1
yarl==1.22.0
zipp==3.23.0
